requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.3.2
distlib==0.3.7
dm-tree==0.1.8
et_xmlfile==2.0.0
filelock==3.12.4
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
pefile==2023.2.7
pillow==11.0.0
pipenv==2023.10.3
platformdirs==3.11.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2023.9
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyVISA==1.14.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scipy==1.12.0
six==1.16.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.24.5
Werkzeug==3.0.1
wrapt==1.14.1
